Marketplace: decimal quantities & lot-backed stock
Convert marketplace quantities to BigDecimal and switch listings to use lot-backed inventory. Added new ProductWarehouseTransactionType entries for marketplace reserve/release and a DB migration V17 for lot-backed inventory. Reworked repositories to query sellable products by lot status/onHandQuantity, added pessimistic locking for lots/products, and new lot-counting queries. Updated DTOs, entities (cart/order item quantities), service logic to reserve/release lot onHandQuantity with ProductWarehouseTransaction records, and various helpers/validators to enforce lot sellability. Tests adjusted accordingly.

+++ agricultural-crop-management-backend/src/main/java/org/example/QuanLyMuaVu/module/marketplace/dto/response/MarketplaceCartResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
 public record MarketplaceCartResponse(
         Long userId,
         List<MarketplaceCartItemResponse> items,
-        Integer itemCount,
+        BigDecimal itemCount,
         BigDecimal subtotal,
         String currency) {
 }
